Split segmentation algorithms into dedicated modules
